backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
orjson>=3.9.10

# HTTP client libraries
aiohttp
requests>=2.31.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
SQLAlchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]>=3.3.0

# Environment and configuration
python-dotenv>=0.21.0

# Caching (removed for Fly.io optimization)
# Email (for password reset functionality)
fastapi-mail>=1.4.1
jinja2>=3.1.2

# Google Gemini AI
google-generativeai